+++ src/auth/auth.service.ts
@@ -112,10 +112,11 @@ export class AuthService {
   async getSessionDetails(cookie: string): Promise<SessionResource> {
     try {
       const userId = await this.getUserId(cookie);
-      const username = await this.getUsername(userId);
+      const { username, avatarUrl } = await this.getProfileData(userId);
       const session: SessionResource = {
         userId,
         username,
+        avatarUrl,
         cookie,
       };
       return session;
@@ -147,11 +148,13 @@ export class AuthService {
   }
 
   /**
-   * Calls the given user's profile page to extract the username.
+   * Calls the given user's profile page to extract the username and avatar.
    * @param userId The user id.
    * @param cookie The session cookie.
    */
-  async getUsername(userId: string): Promise<string> {
+  async getProfileData(
+    userId: string,
+  ): Promise<{ username: string; avatarUrl: string }> {
     const { data } = await this.httpService.get(
       `${forumConfig.USER_PAGE_URL}${userId}`,
     );
@@ -159,6 +162,16 @@ export class AuthService {
     if (!usernameMatches || usernameMatches.length < 3) {
       throw new Error('Unable to retrieve username.');
     }
-    return this.encodingService.decodeText(usernameMatches[2]);
+    const username = this.encodingService.decodeText(
+      usernameMatches[2],
+    ) as string;
+    const avatarUrlMatches = data.match(
+      /(?:(<img\ssrc="\/\/forum.mods.de\/bb\/)(.*)("\sclass="avatar"))/,
+    );
+    if (!avatarUrlMatches || avatarUrlMatches.length < 3) {
+      throw new Error('Unable to retrieve avatar url.');
+    }
+    const avatarUrl = avatarUrlMatches[2];
+    return { username, avatarUrl };
   }
 }