+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [1.1.0] - unreleased
+
+### Added
+
+- Session now includes active avatar URL.
+
+### Fixed
+
+- More fixes and improvements regarding encoding/decoding. Deprecated functions escape() and unescape() are no longer being used.
+- Fixed an issue with the 'updateBoolean' query parameter not working properly on /thread/:id.
+
 ## [1.0.1] - 2023-03-09
 
 ### Fixed